Rename ContactChanged -> ContactRefreshed and EnvironmentChanged -> EnvironmentRefreshed
